+++ packages/react-universal/types/hooks/useActive.d.ts
@@ -1,6 +1,6 @@
-import { HickoryLocation } from "@hickory/root";
+import { SessionLocation } from "@hickory/root";
 import { Params } from "@curi/router";
-export declare type LocationCheck = (l: HickoryLocation) => boolean;
+export declare type LocationCheck = (l: SessionLocation) => boolean;
 export interface ActiveHookProps {
     name: string;
     params?: Params;