+++ packages/router/types/matchLocation.d.ts
@@ -1,4 +1,4 @@
-import { HickoryLocation } from "@hickory/root";
+import { SessionLocation } from "@hickory/root";
 import { CompiledRoute } from "./types/route";
 import { PossibleMatch } from "./types/match";
-export default function matchLocation(location: HickoryLocation, routes: Array<CompiledRoute>): PossibleMatch;
+export default function matchLocation(location: SessionLocation, routes: Array<CompiledRoute>): PossibleMatch;