+++ packages/router/types/types/response.d.ts
@@ -1,4 +1,4 @@
-import { HickoryLocation, PartialLocation } from "@hickory/root";
+import { SessionLocation, PartialLocation } from "@hickory/root";
 import { RouteLocation } from "./location";
 export declare type RawParams = {
     [key: string]: string;
@@ -7,7 +7,7 @@ export declare type Params = {
     [key: string]: any;
 };
 export interface MatchResponseProperties {
-    location: HickoryLocation;
+    location: SessionLocation;
     name: string;
     params: Params;
     partials: Array<string>;