+++ packages/vue/types/Link.d.ts
@@ -1,12 +1,12 @@
 import Vue, { ComponentOptions } from "vue";
-import { HickoryLocation } from "@hickory/root";
+import { SessionLocation } from "@hickory/root";
 export interface LinkComponent extends Vue {
     to: string;
     params?: object;
     hash?: string;
     query?: any;
     state?: any;
-    location: HickoryLocation;
+    location: SessionLocation;
     href: string;
     click(e: MouseEvent): void;
     navigating: boolean;