+++ website/src/pages/Packages/react-dom/v2/contents.js
@@ -4,7 +4,6 @@ import { meta as useCuriMeta } from "./api/useCuri";
 import { meta as useActiveMeta } from "./api/useActive";
 import { meta as useNavigationFocusMeta } from "./api/useNavigationFocus";
 import { meta as useNavigatingMeta } from "./api/useNavigating";
-import { meta as useBlockMeta } from "./api/useBlock";
 import { meta as useLocationMeta } from "./api/useLocation";
 import { meta as useHrefMeta } from "./api/useHref";
 import { meta as CuriousMeta } from "./api/curious";
@@ -28,7 +27,6 @@ export default [
       useActiveMeta,
       useNavigationFocusMeta,
       useNavigatingMeta,
-      useBlockMeta,
       useLocationMeta,
       useHrefMeta,
       CuriousMeta