+++ website/src/pages/Packages/react-native/v2/contents.js
@@ -3,7 +3,6 @@ import { meta as LinkMeta } from "./api/link";
 import { meta as useCuriMeta } from "./api/useCuri";
 import { meta as useActiveMeta } from "./api/useActive";
 import { meta as useNavigatingMeta } from "./api/useNavigating";
-import { meta as useBlockMeta } from "./api/useBlock";
 import { meta as useLocationMeta } from "./api/useLocation";
 import { meta as useHrefMeta } from "./api/useHref";
 import { meta as CuriousMeta } from "./api/curious";
@@ -26,7 +25,6 @@ export default [
       useCuriMeta,
       useActiveMeta,
       useNavigatingMeta,
-      useBlockMeta,
       useLocationMeta,
       useHrefMeta,
       CuriousMeta