+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `navigate` does set default `method`.
+
 ## 2.0.0-alpha.0
 
 * `route.resolve` is now a single function.